// src/typings/yapi/PerpetualOrdersCurrentV1GetApi.d.ts
export interface YapiGetV1PerpetualOrdersCurrentListData {
  /**
   * 订单id
   */
  id: string
  /**
   * 订单类型    limit_order 限价委托单  market_order 市价委托单 forced_liquidation_order 强平委托单 forced_lighten_order 强减委托单
   */
  typeInd: string
  /**
   * 合约组id
   */
  groupId: string
  /**
   * 合约组名称
   */
  groupName: string
  /**
   * 订单状态 unsold 未成交 partially 部分成交
   */
  statusCd: string
  /**
   * 方向 open_long 开多 , open_short 开空 ,close_long 平多, close_short 平空
   */
  sideInd: string
  /**
   * 委托价格类型 limit限价 market市价
   */
  entrustTypeInd: string
  /**
   * 合约交易对id
   */
  tradeId: string
  /**
   * 合约交易对名称
   */
  symbol: string
  /**
   * 杠杆倍数
   */
  lever: number
  /**
   * 委托价格
   */
  price: string
  /**
   * 委托数量
   */
  size: string
  /**
   * 下单选择的币种是标的币还是计价币；如果是标的币传BASE，如果是计价币传QUOTE
   */
  placeUnit: string
  /**
   * 成交均价
   */
  tradePrice: string
  /**
   * 成交数量
   */
  tradeSize: string
  /**
   * 市价单单位 amount按金额 quantity 按数量
   */
  marketUnit: string
  /**
   * 开仓初始仓位保证金
   */
  initMargin: string
  /**
   *  wallet 使用资产账户的资金开仓 group 使用当前合约组的额外保证金开仓
   */
  marginType: string
  /**
   * 开仓额外保证金
   */
  additionalMargin: string
  /**
   *  wallet 使用资产账户的资金开仓 group 使用当前合约组的额外保证金开仓
   */
  additionalMarginType: string
  /**
   * 保证金之和
   */
  totalMargin: string
  /**
   * 完成度
   */
  completeness: string
  /**
   * 创建时间 13位毫秒值
   */
  createdByTime: string
  /**
   * 标的币 symbol
   */
  baseCoinShortName: string
  /**
   * 计价币 symbol
   */
  quoteCoinShortName: string
  /**
   * 是否有止盈止损 true 有 false 无
   */
  hasStrategy: boolean
  /**
   * 使用体验金，开仓使用了体验金优惠券，则该字段有值
   */
  voucherAmount: number
  /**
   * 抵扣的体验金，平仓如果有亏损，如果仓位上有体验金，先扣体验金，则这个字段有值
   */
  voucherDeductionAmount: number
  /**
   * 合约模式：normal 普通模式 taker 带单模式 follower 跟单模式
   */
  modelType: string
  /**
   * 带单跟单模式下的项目ID
   */
  projectId: number
  /**
   * 带单跟单模式下的项目名称
   */
  projectName: string
  /**
   * 带单跟单模式下，分润比例
   */
  shareBonusRate: string
  /**
   * 带单跟单模式下，分润金额
   */
  shareBonusAmount: string
  /**
   * 带单跟单模式下的平仓盈利后，预分润金额
   */
  preDeductionShareBonusAmount: string
}