// src/typings/yapi/PerpetualPlanOrdersCurrentV1GetApi.d.ts
export interface YapiGetV1PerpetualPlanOrdersCurrentListData {
  /**
   * 订单id
   */
  id: string
  /**
   * 合约组id
   */
  groupId: string
  /**
   * 合约组名称
   */
  groupName: string
  /**
   * 委托状态（revoke 已撤销，revoke_sys 系统撤销,not_triggered 未触发 ,alreay_triggered 已触发，triggered_failed 已触发-委托失败）
   */
  statusCd: string
  /**
   * 方向 open_long 开多 , open_short 开空 ,close_long 平多, close_short 平空
   */
  sideInd: string
  /**
   * 委托价格类型 limit限价 market市价
   */
  entrustTypeInd: string
  /**
   * 合约交易对id
   */
  tradeId: string
  /**
   * 合约交易对名称
   */
  symbol: string
  /**
   * 杠杆倍数
   */
  lever: number
  /**
   * 委托价格
   */
  price: string
  /**
   * 委托数量
   */
  size: string
  /**
   * 成交均价
   */
  tradePrice: string
  /**
   * 成交数量
   */
  tradeSize: string
  /**
   * 市价单单位 amount按金额 quantity 按数量
   */
  marketUnit: string
  /**
   * 下单选择的币种是标的币还是计价币；如果是标的币传BASE，如果是计价币传QUOTE
   */
  placeUnit: string
  /**
   * 开仓初始仓位保证金
   */
  initMargin: string
  /**
   *  wallet 使用资产账户的资金开仓 group 使用当前合约组的额外保证金开仓
   */
  marginType: string
  /**
   * 开仓额外保证金
   */
  additionalMargin: string
  /**
   *  wallet 使用资产账户的资金开仓 group 使用当前合约组的额外保证金开仓
   */
  additionalMarginType: string
  /**
   * 保证金之和
   */
  totalMargin: string
  /**
   * 完成度
   */
  completeness: string
  /**
   * 创建时间 13位毫秒值
   */
  createdByTime: string
  /**
   * 标的币 symbol
   */
  baseCoinShortName: string
  /**
   * 计价币 symbol
   */
  quoteCoinShortName: string
  /**
   * 触发价格类型 mark 标记价格 new 最新价格
   */
  triggerPriceTypeInd: string
  /**
   * 触发价格
   */
  triggerPrice: string
  /**
   * 触发方向 up=向上  大于等于，down=向下  小于等于
   */
  triggerDirectionInd: string
  /**
   * 是否有止盈止损   true 有  false没有
   */
  hasStrategy: boolean
  /**
   * 1=专业版 2=普通版，普通版无法添加额外保证金
   */
  perpetualVersion: number
  /**
   * 合约模式：normal 普通模式 taker 带单模式
   */
  modelType: string
  /**
   * 带单模式下的项目ID
   */
  projectId?: string
  /**
   * 带单模式下的项目名称
   */
  projectName?: string
}